% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_list.R
\name{reg_glm}
\alias{reg_glm}
\title{Apply ensemble weighting}
\source{
https://www.tmwr.org/ensembles.html#blend-predictions
https://github.com/tidymodels/stacks/blob/main/R/blend_predictions.R
}
\usage{
reg_glm(
  df_ens,
  target_var,
  group_vars,
  times = 25,
  penalty = 10^(-6:-1),
  mixture = 1
)
}
\arguments{
\item{df_ens}{data.frame}

\item{target_var}{character; column name in df_ens containing the target variable}

\item{group_vars}{character; column names in df_ens to group over}

\item{method}{character;}

\item{...;}{arguments fed into the function specified by 'method'}
}
\description{
Apply user-defined metric and method to weight ensembles
}
\examples{

}
