% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weight_ensemble.R
\name{weight_ensemble}
\alias{weight_ensemble}
\title{Apply ensemble weighting}
\usage{
weight_ensemble(
  df_ens,
  target_var,
  method = "default",
  groups_vars = NULL,
  ...
)
}
\arguments{
\item{df_ens}{data.frame}

\item{target_var}{character; column name in df_ens containing the target variable}

\item{method}{character;}

\item{group_vars}{character; column names in df_ens to group over}

\item{...;}{arguments fed into the function specified by 'method'}
}
\description{
Apply user-defined metric and method to weight ensembles
}
\examples{

}
